[RENAME] 패키지 이름 변경  : notification -> activityNotification
